+++ src/main/drivers/serial_softserial.h
@@ -30,7 +30,7 @@ serialPort_t *softSerialLoopbackPort(void);
 // serialPort API
 void softSerialWriteByte(serialPort_t *instance, uint8_t ch);
 uint32_t softSerialRxBytesWaiting(const serialPort_t *instance);
-uint8_t softSerialTxBytesFree(const serialPort_t *instance);
+uint32_t softSerialTxBytesFree(const serialPort_t *instance);
 uint8_t softSerialReadByte(serialPort_t *instance);
 void softSerialSetBaudRate(serialPort_t *s, uint32_t baudRate);
 bool isSoftSerialTransmitBufferEmpty(const serialPort_t *s);
